render-refactor: complete phase 00 warning-clean Premake baseline
Preserve explicit C++20/static CRT and runtime postbuild/startup checks.
Resolve first-party correctness/type warnings and ShapeConvex duplicate definitions.

One-time owner snapshot waiver: RigidBodySimulation/src/RigidBodySimulation.cpp
Reviewed SHA-256: 8b10d91452c8da64f255aaafcbb4f772da9b2807445253d921d01751afbeef54
Accepted current worktree SHA-256: a8bf14602d46ee7bcbd8a3dd68b6a2d86a04353acc8f038d503700a917f33eb9
Owner quoted prior current SHA-256: 360fed23447df1bebcbc473bb2dcb48d9f8408a6ea7d135587dac28361a49d62
The owner explicitly instructed acceptance of the CURRENT working-tree version.

No Phase 00 builds/tests/runtime validation were rerun for this approval.
Iteration 5 validation is authoritative for all other reviewed files, but does
not fully validate the current committed contents of RigidBodySimulation.cpp.
All other 28 recorded snapshot hashes match. The waiver applies only to the
named file and only to this Phase 00 approval.

Local ignored governance remains untracked. GEngineEditor/imgui.ini is excluded.

+++ GEngine/src/Core/SimpleRenderer.cpp
@@ -14,7 +14,7 @@ namespace GEngine
 
 	namespace Utils
 	{
-		static void FillRGBAToPixel(uint8_t* data, uint32_t index, const Vec4f& pixelColor)
+		static void FillRGBAToPixel(uint8_t* data, size_t index, const Vec4f& pixelColor)
 		{
 			data[4 * index + 0] = (uint8_t)(pixelColor.r * 255.0f);
 			data[4 * index + 1] = (uint8_t)(pixelColor.g * 255.0f);
@@ -151,7 +151,8 @@ namespace GEngine
 				for (size_t x = r.cols().begin(); x < r.cols().end(); x++)
 				{
 						
-					pixelColor = PerPixel(x, y);
+					// These range indices are bounded by the uint32_t image dimensions.
+					pixelColor = PerPixel(static_cast<uint32_t>(x), static_cast<uint32_t>(y));
 					m_AccumulationData[x + y * width] += pixelColor;
 					Vec4f accumulatedColor = m_AccumulationData[x + y * width] / (float)m_FrameIndex;
 					accumulatedColor = glm::clamp(accumulatedColor, Vec4f(0.f), Vec4f(1.0f));
